requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==2.1.1
opencv-python==4.10.0.84
pillow==10.4.0
PyJWT==2.9.0
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.1
wheel==0.44.0